@@ -1,76 +1,83 @@
 {
   "name": "ms-cuentas-digitales",
   "version": "1.0.0",
-  "description": "api intermediaria para el procesamiento de mensajes hacia redis",
+  "description": "Se encarga de administrar las cuentas digitales de Nacional Monte de Piedad",
   "main": "index.js",
   "engines": {
-    "node": ">=14.0.0 <15.0.0",
+    "node": ">=14.14.0 <15.0.0",
     "npm": ">=6.14.4 <7.0.2"
   },
   "scripts": {
     "precise-commits": "precise-commits",
     "pre-commit": "lint-staged",
     "onchange": "onchange",
-    "start": "babel-node src/server/index.js",
+    "test:mongodb": "babel-node ./tests/mongodb.js",
+    "test:ctrl": "mocha --timeout 10000 './tests/**/*.controller.spec.js' --require  @babel/register",
+    "test:service": "mocha --timeout 40000 './tests/**/*.service.spec.js' --require  @babel/register",
+    "test:dao": "mocha --timeout 10000 './tests/**/*.dao.spec.js' --require  @babel/register",
     "test": "nyc --reporter=lcov --reporter=html --reporter=text mocha './tests/**.spec.js' --require  @babel/register --timeout 20000 --exit",
+    "start": "babel-node --tls-min-v1.0 ./src/server",
     "debug": "nodemon --exec babel-node src/server/index.js",
     "eslint": "eslint",
     "lint:base": "eslint src",
     "lint": "npm run lint:base",
     "lint:fix": "eslint --fix --debug src",
     "watch-prettier": "onchange '**/*.jsx' -- prettier --write {{event}} to {{changed}}",
-    "sonar": "node sonar-project.js"
+    "sonar": "node sonar-project.js",
+    "clean": "babel-node ./scripts/clean",
+    "build": "babel-node ./scripts/build",
+    "start:prod": "node ./dist/main.bundle.js"
   },
   "author": "NMP",
   "license": "MIT",
   "dependencies": {
-    "@babel/core": "^7.10.4",
-    "@babel/node": "^7.10.4",
-    "@babel/plugin-transform-runtime": "^7.10.4",
-    "@babel/preset-env": "^7.10.4",
-    "@babel/register": "^7.10.4",
+    "@babel/core": "^7.12.3",
+    "@babel/node": "^7.12.6",
+    "@babel/plugin-transform-runtime": "^7.12.1",
+    "@babel/preset-env": "^7.12.1",
+    "@babel/register": "^7.12.1",
     "babel-plugin-istanbul": "^6.0.0",
-    "base32": "^0.0.6",
-    "base64-async": "^2.1.3",
-    "base64topdf": "^1.1.8",
     "body-parser": "^1.19.0",
-    "cfenv": "^1.2.2",
+    "brakes": "^3.0.1",
+    "cfenv": "^1.2.3",
+    "chalk": "^4.1.0",
     "cors": "^2.8.5",
-    "eslint-config-airbnb-base": "^14.2.0",
+    "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-babel": "^5.3.1",
     "eslint-plugin-import": "^2.22.0",
     "express": "^4.17.1",
-    "express-http-context": "^1.2.4",
-    "ghostscript-node": "^1.3.2",
-    "gm": "^1.23.1",
-    "jsonschema": "^1.2.5",
-    "lodash": "^4.17.19",
+    "ftp": "^0.3.10",
+    "jsonschema": "^1.4.0",
+    "lodash": "^4.17.20",
     "log4js": "^6.3.0",
-    "md5": "^2.3.0",
-    "moment": "^2.27.0",
+    "moment": "^2.29.1",
     "mongoose": "^5.10.15",
     "node-fetch": "^2.6.1",
-    "nodemon": "^2.0.13",
-    "otpauth": "^7.1.3",
-    "redis": "^3.1.2",
-    "sonarqube-scanner": "2.6.0"
+    "nodemon": "^2.0.16",
+    "sonarqube-scanner": "^2.8.0",
+    "stringstream": "^1.0.0",
+    "swagger-jsdoc": "^6.0.0-rc.3",
+    "swagger-ui-express": "^4.1.5"
   },
   "devDependencies": {
+    "babel-loader": "^8.2.2",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
     "chai-http": "^4.3.0",
-    "eslint": "^7.4.0",
-    "eslint-config-prettier": "^6.11.0",
+    "eslint": "^7.13.0",
+    "eslint-config-prettier": "^6.15.0",
     "eslint-plugin-mocha": "^7.0.1",
-    "eslint-plugin-prettier": "^3.1.3",
-    "husky": "^5.1.3",
-    "lint-staged": "^10.2.11",
+    "eslint-plugin-prettier": "^3.1.4",
+    "husky": "^4.2.5",
+    "lint-staged": "^10.5.1",
     "mocha": "^7.2.0",
-    "nock": "^12.0.3",
-    "nyc": "15.0.1",
-    "onchange": "^7.0.2",
     "mongodb-memory-server": "^6.9.2",
+    "nyc": "^15.1.0",
+    "onchange": "^7.1.0",
     "precise-commits": "^1.0.2",
-    "prettier": "^2.0.5"
+    "prettier": "^2.0.5",
+    "progress-bar-webpack-plugin": "^2.1.0",
+    "rimraf": "^3.0.2",
+    "webpack": "^5.11.1"
   }
 }
